Formatting Utilities as Python Project
Utilities is now a properly structured python project. A sample card
template has now acquired, though we will need to come up with something
original eventually
