+++ Final Fantasty/Utilities/card_generation/excelToPNG.py
@@ -1,0 +1,14 @@
+"""This module creates card images for Final Fantasty in a modular fashion using a dictionary that has been
+read from an excel file containing information about the cards. It uses the Pillow library to do some simple text
+and image overlay!
+
+Version 0.1
+
+Author: Rocky Petkov"""
+
+
+from PIL import ImageDraw
+
+
+# Below are some global constants regarding different colours, fonts and file locations to be used with this module.
+# Please do not tamper with these constants with out the express approval of the author.
